Fix bug where CSV write fails if a non-existent initial_header is specified
It is common in the development of a project to want to see a column
near the front of a CSV, but to later remove that column.

Previously, if you didn't remove that column from the
`initial_headers` specification, you'd get an ugly CSV write error
that would lead you down the wrong debugging path.

Now, you get a warning that should encourage you to update your output
file specifications, but the CSV will be written ignoring the missing column.